Silence (spurious) warnings from abstract classes.
Probably just remove the 'doesNotRecognizeSelector:_cmd' implementations,
but remember not to go that route, because it suppresses missing impl
warnings for all concrete subclasses, too, which we don't want (let
subclasses know what they need to implement).

+++ Clojure Lang/CLJAFn.m
@@ -22,7 +22,7 @@
 
 static const inline void CLJAFn_throwArity(NSUInteger actual, NSString *name)
 {
-    NSString *reason = [NSString stringWithFormat:@"Wrong number of args (%lu) passed to: %@", actual, name];
+    NSString *reason = [NSString stringWithFormat:@"Wrong number of args (%lu) passed to: %@", (unsigned long)actual, name];
     @throw [NSException exceptionWithName:NSInvalidArgumentException
                                    reason:reason
                                  userInfo:nil];
